requirements.txt
python-telegram-bot==20.7
telethon==1.33.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
